refactor(pages/workshop): removendo os conflitos - @jcdantss - @gabriele
